AG-16586-csrm-agg-cleanup
